Bug 2045468 - Make sure to check the correct browsing context.  a=RyanVM
Original Revision: https://phabricator.services.mozilla.com/D305611

Differential Revision: https://phabricator.services.mozilla.com/D311831

+++ docshell/base/BrowsingContext.h
@@ -1485,7 +1485,7 @@ class BrowsingContext : public nsILoadContext, public nsWrapperCache {
 
   void DidSet(FieldIndex<IDX_HasSessionHistory>, bool aOldValue);
 
-  bool CanSet(FieldIndex<IDX_BrowserId>, const uint32_t& aValue,
+  bool CanSet(FieldIndex<IDX_BrowserId>, const uint64_t& aValue,
               ContentParent* aSource);
 
   bool CanSet(FieldIndex<IDX_UseErrorPages>, const bool& aUseErrorPages,
